Build #1536 -- Version 0.4.15.xx --------------------------------
* Use 32-/64-bit Size For Certain Functions

Directory trunk:

RESDEBUG.C:
  Use <size_t> as 32-/64-bit size unit for certain functions.

SYMTAB.C:
  Improve comments.
  Use <size_t> as 32-/64-bit size unit for certain functions.
  In IF stmts always compare ptr against NULL and value against 0.

+++ resdebug.c
@@ -1587,7 +1587,7 @@ UBOOL _MyHeapFree
 
 LPVOID _MyVirtualAlloc
     (LPVOID lpAddress,          // Address of region to reserve or commit
-     DWORD  dwSize,             // Size of region
+     size_t dwSize,             // Size of region
      DWORD  flAllocationType,   // Type of allocation
      DWORD  flProtect,          // Type of access protection
      UINT   uLine)              // Line #
